+++ packages/terraform/parser/ObjectStorageBucketParser.ts
@@ -1,5 +1,5 @@
 import { BaseResourceParser } from './BaseResourceParser';
-import { NCloudObjectStorageBucket } from '../model/objectstorage/NCloudObjectStorageBucket';
+import { NCloudObjectStorageBucket } from '../model/objectStorage/NCloudObjectStorageBucket';
 
 export class ObjectStorageBucketParser extends BaseResourceParser {
     protected resourceType = ['object-storage', 'objectstoragebucket'];
