+++ packages/terraform/util/interface/ResourceParsingStrategy.ts
@@ -1,6 +1,6 @@
 import { NCloudModel } from '../../interface/NCloudModel';
 
-export interface ResourceParsingStrategy{
+export interface ResourceParsingStrategy {
     parse(properties: any): NCloudModel;
     canParse(type: string): boolean;
 }
