+++ packages/terraform/util/resourceParser.ts
